RM sub related test
