+++ tests/ttnn/unit_tests/operations/eltwise/test_math.py
@@ -288,3 +288,18 @@ def test_recip_fixed(device, h, w, fill_value):
         torch_output_tensor = golden_function(torch_input_tensor, device=device)
         assert_with_ulp(expected_result=torch_output_tensor, actual_result=output_tensor, ulp_threshold=1)
         assert_allclose(torch_output_tensor, output_tensor, atol=1e-2, rtol=1e-2)
+
+
+@pytest.mark.parametrize("s", [3, 0.5])
+@pytest.mark.parametrize("h", [64])
+@pytest.mark.parametrize("w", [128])
+def test_sub_scalar(device, s, h, w):
+    torch_input_tensor = torch.rand((h, w), dtype=torch.bfloat16)
+    torch_output_tensor = torch_input_tensor - s
+
+    input_tensor = ttnn.from_torch(torch_input_tensor, layout=ttnn.TILE_LAYOUT, device=device)
+
+    output_tensor = input_tensor - s
+    output_tensor = ttnn.to_torch(output_tensor)
+
+    assert_with_ulp(expected_result=torch_output_tensor, actual_result=output_tensor, ulp_threshold=1)
